refactor: move spaces / resources to spa
