+++ apps/admin-ui/src/component/Resources/modules/resourceEditor.ts
