+++ apps/admin-ui/src/component/Resources/queries.ts
