+++ apps/admin-ui/src/component/Spaces/Head.tsx
@@ -5,9 +5,9 @@ import { Link } from "react-router-dom";
 import styled from "styled-components";
 import { breakpoints } from "common/src/common/style";
 import { UnitType } from "common/types/gql-types";
-import { parseAddress } from "../../../common/util";
-import { ContentContainer, IngressContainer } from "../../../styles/layout";
-import LinkPrev from "../../LinkPrev";
+import { parseAddress } from "@common/util";
+import { ContentContainer, IngressContainer } from "@styles/layout";
+import LinkPrev from "@component/LinkPrev";
 
 interface IProps {
   title: string;
