+++ apps/admin-ui/src/component/Spaces/new-space-modal/Page2.tsx
