+++ apps/admin-ui/src/component/Spaces/queries.ts
