+++ apps/admin-ui/src/spa/spaces/[id]/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
 import { useParams } from "react-router-dom";
-import { ContentContainer } from "../../../styles/layout";
-import SpaceEditor from "./SpaceEditor";
+import { ContentContainer } from "@styles/layout";
+import SpaceEditor from "@component/Spaces/SpaceEditor";
 
 type Props = {
   unitPk: string;
